Version 368
http://opensource.apple.com/tarballs/CF/CF-368.tar.gz

+++ PlugIn.subproj/CFPlugIn_Factory.c
@@ -1,9 +1,7 @@
 /*
- * Copyright (c) 2003 Apple Computer, Inc. All rights reserved.
+ * Copyright (c) 2005 Apple Computer, Inc. All rights reserved.
  *
  * @APPLE_LICENSE_HEADER_START@
- * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
  * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
@@ -163,7 +161,7 @@ __private_extern__ void *_CFPFactoryCreateInstance(CFAllocatorRef allocator, _CF
             else {
                 // return values from CFBundleGetFunctionPointerForName will always be dyld, but
                 // we must force-fault them because pointers to glue code do not fault correctly
-                factory->_func = (void *)((unsigned long)(factory->_func) | 0x1);
+                factory->_func = (void *)((uint32_t)(factory->_func) | 0x1);
             }
 #endif
         }
